Changed the file writers into objects whic h generate a hashmap mapping old keys to new keys. Changed a couple of the types of value objects. Might need to change a bunch more when moving ojnto the game specific values (changes from short to int). Made some package refactoring too for readability imp[rovements

+++ src/main/java/org/toby/json/deserialisers/champion/Deserializer.java
@@ -33,7 +33,7 @@ public class Deserializer extends StdDeserializer<Champion> {
         JsonNode titleNode = node.get("title");
         JsonNode rolesNode = node.get("tags");
 
-        champion.setId(idNode.asLong());
+        champion.setId(idNode.asInt());
         champion.setName(nameNode.asText());
         champion.setTitle(titleNode.asText());
         champion.setRoles(serparateString(",", rolesNode.asText()));
